lambda/requirements.txt
goose-extractor
pydub
unidecode
boto3
